almost Final rendition
I was rude last time and didn't put any notes. I was pushing it and I was in a rush. I'll include notes for both revisions below.
created a struct called transaction to help with logic surrounding price history, made a function to read a new file called history.bin which stores transactions, and the supporting function to write to it that is called on transactions. checkPerson has been updated.

in this build, i fixed the bugs I introduced in the last build. see you next commit when invariably i have to fix the bugs introduced by this build!

have a safe day everyone!
(ps no one reads these so :x)